Added VFDEV comments
